instrumentation separation
